Added very basic .osm.pbf file format support (this is a protobuf-based file format, so we added support for protobuf to the project's build_script). We currently only handle DenseNodes and Ways (no regular nodes, Relations, or Changesets). PBF is the file format provided by Geofabrik so we can now open (small) Geofabrik files without needing to convert them to .osm with osmconvert. There's also a lot of cruft code leftover in osm_map.c that really should be cleaned up and moved out

+++ app/osm_map.h
@@ -21,7 +21,7 @@ plex OsmNode
 {
 	u64 id;
 	bool visible;
-	u8 version;
+	i32 version;
 	u64 changeset;
 	Str8 timestampStr;
 	Str8 user;
@@ -43,7 +43,7 @@ plex OsmWay
 {
 	u64 id;
 	bool visible;
-	u8 version;
+	i32 version;
 	u64 changeset;
 	Str8 timestampStr;
 	Str8 user;
